+++ libs/docs/core/popover/popover-header/popover-header.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { IconComponent } from '@fundamental-ngx/core/icon';
 import { MessageStripComponent } from '@fundamental-ngx/core/message-strip';
 import {
     DescriptionComponent,
@@ -19,7 +20,8 @@ import {
         DescriptionComponent,
         MessageStripComponent,
         ImportComponent,
-        HeaderTabsComponent
+        HeaderTabsComponent,
+        IconComponent
     ]
 })
 export class PopoverHeaderComponent {}