+++ libs/docs/core/side-navigation/side-navigation-header/side-navigation-header.component.ts
@@ -1,4 +1,7 @@
 import { Component } from '@angular/core';
+import { RouterLink } from '@angular/router';
+import { LinkComponent } from '@fundamental-ngx/core/link';
+import { MessageStripComponent } from '@fundamental-ngx/core/message-strip';
 import {
     DescriptionComponent,
     DocPageComponent,
@@ -12,6 +15,15 @@ import {
     templateUrl: './side-navigation-header.component.html',
     styleUrls: ['./side-navigation-header.component.scss'],
     standalone: true,
-    imports: [DocPageComponent, HeaderComponent, DescriptionComponent, ImportComponent, HeaderTabsComponent]
+    imports: [
+        DocPageComponent,
+        HeaderComponent,
+        DescriptionComponent,
+        ImportComponent,
+        HeaderTabsComponent,
+        MessageStripComponent,
+        RouterLink,
+        LinkComponent
+    ]
 })
 export class SideNavigationHeaderComponent {}