+++ libs/docs/core/tabs/examples/tab-form-element-example/tab-form-element-example.component.ts
@@ -3,11 +3,12 @@ import { ButtonComponent } from '@fundamental-ngx/core/button';
 import { FormItemComponent, FormLabelComponent } from '@fundamental-ngx/core/form';
 import { InputGroupModule } from '@fundamental-ngx/core/input-group';
 import { TabsModule } from '@fundamental-ngx/core/tabs';
+import { TextComponent } from '@fundamental-ngx/core/text';
 
 @Component({
     selector: 'fd-tab-form-element-example',
     templateUrl: './tab-form-element-example.component.html',
     standalone: true,
-    imports: [FormLabelComponent, TabsModule, ButtonComponent, FormItemComponent, InputGroupModule]
+    imports: [FormLabelComponent, TabsModule, ButtonComponent, FormItemComponent, InputGroupModule, TextComponent]
 })
 export class TabFormElementExampleComponent {}