+++ libs/docs/shared/src/lib/core-helpers/stackblitz/stackblitz-dependencies.ts
@@ -5,7 +5,8 @@ export class StackblitzDependencies {
         '@fundamental-ngx/core',
         '@fundamental-ngx/i18n',
         '@fundamental-ngx/cx',
-        '@fundamental-ngx/btp'
+        '@fundamental-ngx/btp',
+        '@fundamental-ngx/datetime-adapter'
     ];
 
     private static _ngDependencies: string[] = [
@@ -29,6 +30,7 @@ export class StackblitzDependencies {
         'core-js',
         'focus-trap',
         'fundamental-styles',
+        '@sap-ui/common-css',
         'dayjs',
         'moment',
         'tslib',