+++ libs/cx/CHANGELOG.md
@@ -3,6 +3,246 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.50.0-rc.34](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.33...v0.50.0-rc.34) (2024-04-17)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.33](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.32...v0.50.0-rc.33) (2024-04-17)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.32](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.31...v0.50.0-rc.32) (2024-04-16)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.31](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.30...v0.50.0-rc.31) (2024-04-16)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.30](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.29...v0.50.0-rc.30) (2024-04-16)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.29](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.28...v0.50.0-rc.29) (2024-04-16)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.28](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.27...v0.50.0-rc.28) (2024-04-16)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.27](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.26...v0.50.0-rc.27) (2024-04-16)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.26](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.25...v0.50.0-rc.26) (2024-04-16)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.25](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.24...v0.50.0-rc.25) (2024-04-12)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.24](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.23...v0.50.0-rc.24) (2024-04-11)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.23](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.22...v0.50.0-rc.23) (2024-04-11)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.22](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.19...v0.50.0-rc.22) (2024-04-11)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.21](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.19...v0.50.0-rc.21) (2024-04-11)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.19](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.18...v0.50.0-rc.19) (2024-04-11)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.18](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.17...v0.50.0-rc.18) (2024-04-11)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.17](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.16...v0.50.0-rc.17) (2024-04-11)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.16](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.15...v0.50.0-rc.16) (2024-04-09)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.15](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.14...v0.50.0-rc.15) (2024-04-09)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.14](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.13...v0.50.0-rc.14) (2024-04-09)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.13](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.12...v0.50.0-rc.13) (2024-04-08)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.12](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.11...v0.50.0-rc.12) (2024-04-08)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.11](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.10...v0.50.0-rc.11) (2024-04-08)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.10](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.9...v0.50.0-rc.10) (2024-04-07)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.9](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.8...v0.50.0-rc.9) (2024-04-07)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.8](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.7...v0.50.0-rc.8) (2024-04-07)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.7](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.6...v0.50.0-rc.7) (2024-04-07)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.6](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.5...v0.50.0-rc.6) (2024-04-06)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.5](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.4...v0.50.0-rc.5) (2024-04-06)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
+# [0.50.0-rc.4](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.3...v0.50.0-rc.4) (2024-04-05)
+
+**Note:** Version bump only for package @fundamental-ngx/cx
+
+
+
+
+
 # [0.50.0-rc.3](https://github.com/SAP/fundamental-ngx/compare/v0.50.0-rc.2...v0.50.0-rc.3) (2024-04-04)
 
 **Note:** Version bump only for package @fundamental-ngx/cx